Implemented tabs and save bar visual
