+++ package.json
@@ -50,7 +50,7 @@
         "lucide-react": "^0.545.0",
         "maplibre-gl": "^5.16.0",
         "next": "^16.0.7",
-        "nuqs": "^2.8.6",
+        "nuqs": "^2.8.9",
         "pg": "^8.16.3",
         "react": "^19.2.2",
         "react-dom": "^19.2.2",
