+++ src/app/settings/page.tsx
@@ -2,7 +2,7 @@
  *
  *                page.tsx
  *
- *         Author: Will and Hansini
+ *         Author: Will, Hansini, and Justin
  *           Date: 12/6/2025
  *
  *        Summary: Basic outline of settings page
@@ -12,7 +12,7 @@
 "use client";
 
 import { useCallback, useEffect, useState } from "react";
-import { Trash, Plus, Pencil } from "lucide-react";
+import { Trash, Pencil } from "lucide-react";
 import { Button } from "@/components/ui/button";
 import { Combobox } from "@/components/Combobox";
 import YearsOfData from "@/components/YearsOfData";
@@ -20,236 +20,97 @@ import { Map, MapMarker, MarkerContent, useMap } from "@/components/ui/map";
 import { toast } from "sonner";
 import GatewaySchools from "@/components/GatewaySchools";
 import { standardize } from "@/lib/school-name-standardize";
-
-interface PermittedUser {
-    email: string;
-    lastSignIn: string;
-}
+import { Tabs, TabsList, TabsTrigger, TabsContent } from "@/components/ui/tabs";
+import PermittedUsers from "@/components/ui/PermittedUsers";
 
 export default function Settings() {
-    const [emailInput, setEmailInput] = useState("");
-    const [permittedUsers, setPermittedUsers] = useState<PermittedUser[]>([
-        {
-            email: "something@mhd.com",
-            lastSignIn: "3 days ago",
-        },
-        {
-            email: "something@mhd.com",
-            lastSignIn: "3 days ago",
-        },
-        {
-            email: "something@mhd.com",
-            lastSignIn: "3 days ago",
-        },
-    ]);
-
     const [hasUnsavedChanges, setHasUnsavedChanges] = useState(false);
 
     const handleSave = () => {
         setHasUnsavedChanges(false);
     };
 
-    // Email validation function
-    const isValidEmail = (email: string): boolean => {
-        const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-        return emailRegex.test(email);
-    };
-
-    // Add permitted user
-    const handleAddUser = () => {
-        const trimmedEmail = emailInput.trim();
-
-        // Validate email format
-        if (!isValidEmail(trimmedEmail)) {
-            alert("Please enter a valid email address");
-            return;
-        }
-
-        // Check for duplicates
-        if (
-            permittedUsers.some(
-                (user) =>
-                    user.email.toLowerCase() === trimmedEmail.toLowerCase(),
-            )
-        ) {
-            alert("This email is already in the permitted users list");
-            return;
-        }
-
-        // Add new user
-        setPermittedUsers([
-            ...permittedUsers,
-            {
-                email: trimmedEmail,
-                lastSignIn: "3 days ago", // Mock data as requested
-            },
-        ]);
-
-        // Clear input
-        setEmailInput("");
-        setHasUnsavedChanges(true);
-    };
-
-    // Remove permitted user
-    const handleRemoveUser = (email: string) => {
-        setPermittedUsers(
-            permittedUsers.filter((user) => user.email !== email),
-        );
-        setHasUnsavedChanges(true);
-    };
-
-    // Handle Enter key press
-    const handleEmailKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
-        if (e.key === "Enter") {
-            handleAddUser();
-        }
+    const handleDiscard = () => {
+        setHasUnsavedChanges(false);
     };
 
     return (
-        <div className="flex flex-col gap-12 p-6 max-w-4xl">
+        <div className="flex flex-col gap-12 p-6 max-w-4xl pb-24">
             <div>
                 <h1 className="text-3xl font-bold">Settings</h1>
             </div>
 
-            <section className="space-y-6">
-                <h2 className="text-xl font-semibold">Preferences</h2>
-                <p className="text-gray-600">
-                    How would you like to view charts...
-                </p>
-            </section>
+            {/*Where does this go? */}
+            <h2 className="text-xl font-semibold">Preferences</h2>
+            <p className="text-gray-600">
+                How would you like to view charts...
+            </p>
 
-            <section className="space-y-6">
-                <div>
-                    <h2 className="text-xl font-semibold">Configuration</h2>
-                    <p className="text-gray-600">
-                        These settings configure how data is calculated. Only
-                        edit these settings if you really mean to.
-                    </p>
-                </div>
+            <Tabs defaultValue="data-management">
+                <TabsList>
+                    <TabsTrigger value="data-management">
+                        Data Management
+                    </TabsTrigger>
+                    <TabsTrigger value="team-access">Team & Access</TabsTrigger>
+                </TabsList>
 
-                <div className="mt-6 space-y-6">
-                    <div className="space-y-3">
-                        <div className="flex items-center justify-between">
-                            <h3 className="font-bold">Gateway Schools</h3>
-                        </div>
-                        <GatewaySchools />
-                    </div>
-                    {/* School Locations Section */}
-                    <SchoolLocationEditor />
-
-                    {/* Permitted Users Section */}
-                    <div className="space-y-3">
-                        <div className="flex items-center justify-between">
-                            <h3 className="font-bold">Permitted Users</h3>
-                        </div>
-
-                        <p className="text-sm text-gray-600">
-                            These emails are permitted to sign into the
-                            platform. Here you can also revoke access
+                <TabsContent value="data-management" className="mt-6 space-y-6">
+                    <div>
+                        <h2 className="text-xl font-semibold">Configuration</h2>
+                        <p className="text-gray-600">
+                            These settings configure how data is calculated.
+                            Only edit these settings if you really mean to.
                         </p>
-                        <div className="flex rounded-lg border border-gray-300 shadow-sm overflow-hidden w-72">
-                            <input
-                                type="email"
-                                placeholder="Email"
-                                aria-label="Email"
-                                value={emailInput}
-                                onChange={(e) => setEmailInput(e.target.value)}
-                                onKeyDown={handleEmailKeyDown}
-                                className="flex-1 px-4 py-1 text-base text-gray-700 placeholder-gray-500 outline-none"
-                            />
-                            <button
-                                type="button"
-                                onClick={handleAddUser}
-                                aria-label="Add Email"
-                                className="bg-gray-200 hover:bg-gray-300 text-gray-600 transition-colors w-8 flex items-center justify-center border-l border-gray-300"
-                            >
-                                <Plus className="w-6 h-6" />
-                            </button>
-                        </div>
-
-                        <div className="border border-gray-200 rounded-lg overflow-hidden">
-                            <table className="w-full">
-                                <thead className="bg-gray-50">
-                                    <tr>
-                                        <th className="text-left px-4 py-3 text-sm font-medium text-gray-700">
-                                            Email
-                                        </th>
-                                        <th className="text-left px-4 py-3 text-sm font-medium text-gray-700">
-                                            Last Signed In
-                                        </th>
-                                        <th className="text-left px-4 py-3 text-sm font-medium text-gray-700">
-                                            Actions
-                                        </th>
-                                    </tr>
-                                </thead>
-                                <tbody className="divide-y divide-gray-200 bg-white">
-                                    {permittedUsers.map((user, i) => (
-                                        <tr
-                                            key={i}
-                                            className="hover:bg-gray-50"
-                                        >
-                                            <td className="px-4 py-3 text-sm">
-                                                {user.email}
-                                            </td>
-                                            <td className="px-4 py-3 text-sm text-gray-600">
-                                                {user.lastSignIn}
-                                            </td>
-                                            <td className="p-4">
-                                                <button
-                                                    onClick={() =>
-                                                        handleRemoveUser(
-                                                            user.email,
-                                                        )
-                                                    }
-                                                    className="text-gray-400 hover:text-red-500 transition-colors"
-                                                    aria-label={`Remove ${user.email}`}
-                                                >
-                                                    <Trash className="w-4 h-4" />
-                                                </button>
-                                            </td>
-                                        </tr>
-                                    ))}
-                                </tbody>
-                            </table>
-                        </div>
                     </div>
-
-                    <div className="mt-6 space-y-6">
+                    <div className="space-y-6">
                         <div className="space-y-3">
-                            <div className="space-y-6 w-full">
-                                <h3 className="font-bold">Available Data</h3>
-                                <YearsOfData />
-                            </div>
+                            <h3 className="font-bold">Gateway Cities</h3>
+                            <GatewaySchools
+                                onUnsavedChange={() =>
+                                    setHasUnsavedChanges(true)
+                                }
+                            />
+                        </div>
+                        <SchoolLocationEditor />
+                        <div className="space-y-3">
+                            <h3 className="font-bold">Available Data</h3>
+                            <YearsOfData
+                                onUnsavedChange={() =>
+                                    setHasUnsavedChanges(true)
+                                }
+                            />
                         </div>
                     </div>
-                </div>
+                </TabsContent>
 
-                {/* Save Section */}
-                <div className="flex items-center justify-between py-6">
-                    <label className="flex items-center gap-2 text-sm text-gray-600">
-                        <span>You have unsaved changes - save?</span>
-                    </label>
+                <TabsContent value="team-access" className="mt-6 space-y-5">
+                    <PermittedUsers
+                        onUnsavedChange={() => setHasUnsavedChanges(true)}
+                    />
+                </TabsContent>
+            </Tabs>
+
+            {hasUnsavedChanges && (
+                <div className="fixed bottom-0 left-56 right-0 z-50 flex items-center justify-between px-8 py-4 bg-white border-t border-gray-200 shadow-lg">
+                    <span className="text-sm text-gray-600">
+                        You have unsaved changes
+                    </span>
                     <div className="flex gap-2">
                         <button
-                            className="pl-2 pr-6 py-2 text-base text-gray-700 placeholder-gray-500
-                            rounded-lg border border-gray-300 shadow-sm w-6"
+                            onClick={handleDiscard}
+                            className="pl-2 pr-2 py-2 text-base text-gray-700 rounded-lg border border-gray-300 shadow-sm flex items-center gap-2 hover:bg-gray-100"
                         >
                             <Trash className="h-4 w-4" />
                         </button>
                         <button
                             onClick={handleSave}
-                            disabled={!hasUnsavedChanges}
-                            className={`px-6 py-2 rounded-lg ${
-                                hasUnsavedChanges
-                                    ? "bg-blue-600 text-white hover:bg-blue-700"
-                                    : "bg-gray-200 text-gray-400 cursor-not-allowed"
-                            }`}
+                            className="px-6 py-2 rounded-lg bg-blue-600 text-white hover:bg-blue-700"
                         >
                             Save
                         </button>
                     </div>
                 </div>
-            </section>
+            )}
         </div>
     );
 }
@@ -390,9 +251,6 @@ function SchoolLocationEditor() {
             <div className="flex items-center justify-between">
                 <h3 className="font-bold">School Locations</h3>
             </div>
-            <p className="text-sm text-gray-600">
-                Search for a school to view and edit its location on the map.
-            </p>
             <div className="min-w-72 w-fit">
                 <Combobox
                     options={schoolOptions}
